fix(ZMSKVR-1453): fix service info link in rebooking and appointment overview
Extract resolveServiceInfoId and use it when loading appointments for
rebooking/overview so nested variants link to the root service instead
of the immediate parent.
